refactor(age-review): hoist the enforcement/response contract types into shared
EnforcementLegStatus, AgeReviewEnforcement, and the case-response shape were
defined twice -- inline in the worker (worker/src/age-review.ts) and again in
the admin client (src/lib/adminApi.ts). Move the canonical definitions into
shared/age-review.ts so the worker API and client share one source of truth and
cannot drift. Type-only, no behavior change.

Surfaced in the #122 review.
